Time: 770 ms (5.38%), Space: 100.4 MB (91.94%) - LeetHub

+++ 3075-maximize-happiness-of-selected-children/3075-maximize-happiness-of-selected-children.java
@@ -2,7 +2,7 @@ class Solution {
   
     public long maximumHappinessSum(int[] happiness, int k) {
         Arrays.sort(happiness);
-        int max = 0;
+        long max = 0;
         int idx = happiness.length - 1;
         int decrement = 1;
         while(k > 0){
